+++ CHANGELOG.md
@@ -7,7 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 * Breaking change: you should remove dependency in "kotless-lang-aws" as its no longer part of the project
 * removed ktor & kotless dsl
 * removed azure support
-* support sns consumption
+* support sns
+* support gamelift permission
 
 # 0.3.3 - 2023-12-25